requirements/requirements-testing.txt
-r requirements-base.txt

# PyTest for running the tests.
pytest
pytest-cov
pytest-django

flake8
